requirements.txt
fastapi
uvicorn[standard]
twilio
openai
python-dotenv
pydantic
pydantic-settings
python-multipart
httpx
redis
